*sigh*
Be sure to type your variables in the correct case. The instances are meant to be created using curly braces, please don't change that again. Anyone who wants to add to this should do their editing in eclipse, and be sure to install the necessary toolchains Also, please look at the documentation for the tolower method. You're using it wrong